tests: Check saved positions without expanding selected ranges
The merge code has a function that inspects saved insertion, removal, and
replacement records without enumerating every selected line. Edit execution
still contains the former copy.

The large, fragmented-selection test calls the former location. Production
cannot delete that copy first because source and test changes remain in
separate commits.

This commit continues the move by placing the test beside the new function
and calling that function directly.

The next source commit will use the tested function when choosing how to
locate changes and remove the copy from edit execution.
